% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addLastValueToList}
\alias{addLastValueToList}
\title{Add Last Value to Cache List}
\usage{
addLastValueToList(value)
}
\arguments{
\item{value}{The value to add to the cache list.}
}
\value{
The updated cache list of last values (invisibly). This
  function is typically called for its side effect of updating the cache list
  and normally does not need to return a value.
}
\description{
Adds the given value to a globally accessible list which is meant to store the last evaluated values.
The function ensures the list contains a user-defined maximum number of elements, discarding the oldest value if necessary.
}
