% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_values.R
\name{getMaxCachedValues}
\alias{getMaxCachedValues}
\title{Get Maximum Number of Cached Values}
\usage{
getMaxCachedValues()
}
\value{
The currently set maximum number of last values to cache.
}
\description{
Retrieves the currently set maximum number of last values to cache defined by the user.
}
