% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setMaxCachedValues}
\alias{setMaxCachedValues}
\title{Set Maximum Number of Cached Values}
\usage{
setMaxCachedValues(num)
}
\arguments{
\item{n}{Numerical value indicating the maximum number of last values to maintain in the cache.}
}
\value{
Invisible `NULL`. The function is used for its side effect of setting
  an option and typically does not need to return a value.
}
\description{
Sets the maximum number of last values to cache.
This function changes the option that controls the maximum number of cached values allowed.
}
